move tests to make sure internal functionality not used
